Some changes, for now.
Just "finished" this site. I think it is ok to end like this. However, i know that i could make more in this site. But, as the main purpose of it is to learn somethings about html and css, i think it's enough to stop here, for now, and keep my lessons with another challengs and others sites to build.  Any tips, just contact. Thx.